Rename Tab to GenericTabs and Collapsible to GenericCollapsible (#353)
* Rename Tab to GenericTabs

* Rename Collapsible to GenericCollapsible

* Update snapshots
